Move logger to its own subfolder
